refactor(runtime): change token_multipliers from int penalties to float multipliers
Replace the 0-100 integer penalty format with a direct 0.0-1.0 multiplier float so
callers no longer need to convert. Adds a config migration for existing TOML files.

+++ src/subsearch/runtime/models/model.py
@@ -56,7 +56,7 @@ class AppConfig:
     only_foreign_parts: bool
     providers: dict[str, bool]
     token_weights: dict[str, float]
-    token_multipliers: dict[str, int]
+    token_multipliers: dict[str, float]
     context_menu: bool
     context_menu_icon: bool
     file_extensions: dict[str, bool]
